Split tests in different files according to functionaility
