+++ my_app/tests/test_user_model.py
@@ -1,27 +1,10 @@
-import json, flask
-from my_app import db, app, bcrypt
+from .test_base import TestBase
+import json
 from my_app.product.models import User, Bucketlist
-import unittest
+from my_app import db, app, bcrypt
 
 
-class TestUserModel(unittest.TestCase):
-    def setUp(self):
-        app.config.from_object('config.TestingConfigs')
-        db.session.close()
-        db.drop_all()
-        db.create_all()
-        user = User('testi@mail.com', bcrypt.generate_password_hash('testpass').decode('utf-8'))
-        db.session.add(user)
-        db.session.commit()
-        bucketlist = Bucketlist(name='Cities')
-        db.session.add(bucketlist)
-        user.bucketlists.append(bucketlist)  # FK relationship
-        db.session.commit()
-        self.bucketlistID = bucketlist.id
-        self.auth_token = user.encode_auth_token(user.id)
-        self.app = app.test_client()
-        self.user = user
-        return self.app, self.auth_token, self.user, self.bucketlistID
+class TestUserModel(TestBase):
 
     def test_error_handler(self):
         rv1 = self.app.get('/api/v1/bucketlis/', headers=dict(Authorization=self.auth_token))
@@ -148,390 +131,3 @@ class TestUserModel(unittest.TestCase):
         rv = self.app.post('/api/v1/auth/logout')
         data = json.loads(rv.data.decode())
         self.assertEqual('Token not found, Login to get one', data['message'])
-
-    def create_bucketlist(self, bucketlist_name, token):
-        return self.app.post('/api/v1/bucketlists/', data=dict(
-            name=bucketlist_name), headers=dict(Authorization=token))
-
-    def test_user_can_add_bucketlist(self):
-        rv = self.create_bucketlist('Oceans', self.auth_token)
-        data = json.loads(rv.data.decode())
-        self.assertEqual('Oceans', data['bucketlist'])
-
-    def test_user_cannot_add_bucketlist_without_name(self):
-        rv = self.create_bucketlist('', self.auth_token)
-        data = json.loads(rv.data.decode())
-        self.assertEqual('Bucketlist item should have name', data['message'])
-
-    def test_user_cannot_add_already_existing_bucketlist(self):
-        rv = self.create_bucketlist('Cities', self.auth_token)
-        data = json.loads(rv.data.decode())
-        self.assertEqual('Bucketlist exists',data['message'])
-
-    def test_unauthenticated_user_cannot_add_bucketlist(self):
-        rv = self.create_bucketlist('Cities', None)
-        data = json.loads(rv.data.decode())
-        self.assertEqual('Invalid token, Login again',data['message'])
-
-    def test_return_error_when_token_is_not_found(self):
-        rv = self.app.post('/api/v1/bucketlists/', data=dict(name='Work'), headers=dict())
-        data = json.loads(rv.data.decode())
-        self.assertEqual('Token not found, Login to get one',data['message'])
-
-    def view_all_bucketlists(self, token):
-        return self.app.get('/api/v1/bucketlists/', headers=dict(
-            Authorization=token
-        ))
-
-    def test_user_can_view_bucketlists(self):
-        rv = self.view_all_bucketlists(self.auth_token)
-        data = json.loads(rv.data.decode())
-        self.assertEqual(data['1'], 'Cities')
-
-    def test_unauthenticated_user_cannot_view_bucketlist(self):
-        rv = self.view_all_bucketlists(None)
-        data = json.loads(rv.data.decode())
-        self.assertEqual('Invalid token, Login again',data['message'])
-
-    with app.test_client() as c:
-        rv = c.get('/api/v1/bucketlists/?q=Cities')
-        assert flask.request.args['q'] == 'Cities'
-        assert flask.request.path == '/api/v1/bucketlists/'
-
-    def test_pagination(self):
-        rv = self.app.get('/api/v1/bucketlists/?limit=2', headers=dict(
-            Authorization=self.auth_token
-        ))
-        data = json.loads(rv.data.decode())
-        self.assertEqual(1, len(data))
-
-    def test_search_and_pagination_of_bucketlist(self):
-        self.create_bucketlist('Oceans', self.auth_token)
-        self.create_bucketlist('Oceania', self.auth_token)
-        self.create_bucketlist('Octopie', self.auth_token)
-        rv = self.app.get('/api/v1/bucketlists/?q=Oc&limit=2', headers=dict(
-            Authorization=self.auth_token
-        ))
-        data = json.loads(rv.data.decode())
-        self.assertEqual(len(data), 2)
-        rv = self.app.get('/api/v1/bucketlists/?q=Zanza&limit=3', headers=dict(
-            Authorization=self.auth_token
-        ))
-        data = json.loads(rv.data.decode())
-        self.assertEqual(data['message'], 'No result matches this search')
-
-    def test_name_based_search_of_bucketlist(self):
-        rv = self.app.get('/api/v1/bucketlists/?q=Cities', headers=dict(
-            Authorization=self.auth_token
-        ))
-        data = json.loads(rv.data.decode())
-        self.assertTrue(data['1'])
-
-    # def test_get_existing_bucketlist(self):
-    #     rv = self.app.get('/bucketlists/1', headers=dict(
-    #         Authorization=self.auth_token
-    #     ))
-    #     data = json.loads(rv.data.decode())
-    #     print (data)
-    #     #self.assertFalse(data['1'])
-
-    def test_authenticated_user_can_update_bucketlist(self):
-        bucketlist = Bucketlist(name='Career')
-        db.session.add(bucketlist)
-        self.user.bucketlists.append(bucketlist)  # FK relationship
-        db.session.commit()
-        bucketlistID = bucketlist.id
-        rv = self.app.put('/api/v1/bucketlists/'+ str(bucketlistID), data=json.dumps(dict(newname='Work')),
-                          headers=dict(Authorization=self.auth_token))
-        data = json.loads(rv.data.decode())
-        self.assertEqual(data['message'], 'Bucketlist updated successfully')
-
-    def test_authenticated_user_cannot_update_bucketlist_that_doesnt_exist(self):
-        rv = self.app.put('/api/v1/bucketlists/'+ str(200), data=json.dumps(dict(newname='Work')),
-                          headers=dict(Authorization=self.auth_token))
-        data = json.loads(rv.data.decode())
-        self.assertEqual(data['message'], 'Bucketlist doesnt exist')
-
-    def test_unauthenticated_user_cannot_update_bucketlist(self):
-        bucketlist = Bucketlist(name='Career')
-        db.session.add(bucketlist)
-        self.user.bucketlists.append(bucketlist)  # FK relationship
-        db.session.commit()
-        bucketlistID = bucketlist.id
-        rv = self.app.put('/api/v1/bucketlists/'+ str(bucketlistID), data=json.dumps(dict(newname='Work')),
-                          headers=dict(Authorization=None))
-        data = json.loads(rv.data.decode())
-        self.assertEqual(data['message'], 'Invalid token, Login again')
-
-    def test_user_cannot_update_bucketlist_without_token(self):
-        bucketlist = Bucketlist(name='Career')
-        db.session.add(bucketlist)
-        self.user.bucketlists.append(bucketlist)  # FK relationship
-        db.session.commit()
-        bucketlistID = bucketlist.id
-        rv = self.app.put('/api/v1/bucketlists/'+ str(bucketlistID), data=json.dumps(dict(newname='Work')))
-        data = json.loads(rv.data.decode())
-        self.assertEqual(data['message'], 'Token not found, Login to get one')
-
-    def test_user_can_delete_bucketlist(self):
-        bucketlist = Bucketlist(name='Visit')
-        db.session.add(bucketlist)
-        self.user.bucketlists.append(bucketlist)  # FK relationship
-        db.session.commit()
-        bucketlistID = bucketlist.id
-        rv = self.app.delete('/api/v1/bucketlists/'+ str(bucketlistID), headers=dict(Authorization=self.auth_token))
-        data = json.loads(rv.data.decode())
-        self.assertEqual(data['message'], 'Bucketlist deleted successfully')
-
-    def test_user_cannot_delete_bucketlist_that_doesnt_exist(self):
-        rv = self.app.delete('/api/v1/bucketlists/'+ str(100), headers=dict(Authorization=self.auth_token))
-        data = json.loads(rv.data.decode())
-        self.assertEqual(data['message'], 'Bucketlist doesnt exist')
-
-    def test_unauthenticated_user_cannot_delete_bucketlist(self):
-        bucketlist = Bucketlist(name='Visit')
-        db.session.add(bucketlist)
-        self.user.bucketlists.append(bucketlist)  # FK relationship
-        db.session.commit()
-        bucketlistID = bucketlist.id
-        rv = self.app.delete('/api/v1/bucketlists/'+ str(bucketlistID), headers=dict(Authorization=None))
-        data = json.loads(rv.data.decode())
-        self.assertEqual(data['message'], 'Invalid token, Login again')
-
-    def test_user_cannot_delete_bucketlist_without_token(self):
-        bucketlist = Bucketlist(name='Visit')
-        db.session.add(bucketlist)
-        self.user.bucketlists.append(bucketlist)  # FK relationship
-        db.session.commit()
-        bucketlistID = bucketlist.id
-        rv = self.app.delete('/api/v1/bucketlists/'+ str(bucketlistID))
-        data = json.loads(rv.data.decode())
-        self.assertEqual(data['message'], 'Token not found, Login to get one')
-
-    def test_authenticated_user_can_add_item_to_bucketlist(self):
-        bucketlist = Bucketlist(name='Water bodies')
-        db.session.add(bucketlist)
-        self.user.bucketlists.append(bucketlist)  # FK relationship
-        db.session.commit()
-        bucketlistID = bucketlist.id
-        rv = self.app.post('/api/v1/bucketlists/' + str(bucketlistID)+ '/items/', data=dict(
-            title='Indian Ocean',
-            deadline='13/11/1994',
-            status='Incomplete'
-        ), headers=dict(Authorization=self.auth_token))
-        data = json.loads(rv.data.decode())
-        self.assertEqual(data['message'], 'Bucketlist item added successfully')
-        # rv = self.app.get('/api/v1/bucketlists/'+ str(bucketlistID), headers=dict(
-        #     Authorization=self.auth_token
-        # ))
-        # data = json.loads(rv.data.decode())
-        # print(data)
-        # self.assertFalse(data['1'])
-
-    def test_authenticated_user_cannot_add_item_to_bucketlist_that_doesnt_exist(self):
-        rv = self.app.post('/api/v1/bucketlists/' + str(100) + '/items/', data=dict(
-            title='Indian Ocean',
-            deadline='13/11/1994',
-            status='Incomplete'
-        ), headers=dict(Authorization=self.auth_token))
-        data = json.loads(rv.data.decode())
-        # print(data['message'])
-        self.assertEqual(data['message'], "Bucketlist doesn't exist")
-
-    def test_unauthenticated_user_cannot_add_item_to_bucketlist(self):
-        bucketlist = Bucketlist(name='Water bodies')
-        db.session.add(bucketlist)
-        self.user.bucketlists.append(bucketlist)  # FK relationship
-        db.session.commit()
-        bucketlistID = bucketlist.id
-        rv = self.app.post('/api/v1/bucketlists/' + str(bucketlistID) + '/items/', data=dict(
-            title='Indian Ocean',
-            deadline='13/11/1994',
-            status='Incomplete'
-        ), headers=dict(Authorization=None))
-        data = json.loads(rv.data.decode())
-        # print(data['message'])
-        self.assertEqual(data['message'], 'Invalid token, Login again')
-
-    def test_user_cannot_add_item_to_bucketlist_without_token(self):
-        bucketlist = Bucketlist(name='Water bodies')
-        db.session.add(bucketlist)
-        self.user.bucketlists.append(bucketlist)  # FK relationship
-        db.session.commit()
-        bucketlistID = bucketlist.id
-        rv = self.app.post('/api/v1/bucketlists/' + str(bucketlistID) + '/items/', data=dict(
-            title='Indian Ocean',
-            deadline='13/11/1994',
-            status='Incomplete'
-        ))
-        data = json.loads(rv.data.decode())
-        # print(data['message'])
-        self.assertEqual(data['message'], 'Token not found, Login to get one')
-
-    def test_authenticated_user_can_get_item_in_bucketlist(self):
-        bucketlist = Bucketlist(name='Water bodies')
-        db.session.add(bucketlist)
-        self.user.bucketlists.append(bucketlist)  # FK relationship
-        db.session.commit()
-        bucketlistID = bucketlist.id
-
-        rv1 = self.app.get('/api/v1/bucketlists/' + str(bucketlistID), headers=dict(Authorization=self.auth_token))
-        data1 = json.loads(rv1.data.decode())
-        self.assertEqual(data1['message'], 'No items in this bucketlist')
-        self.app.post('/api/v1/bucketlists/' + str(bucketlistID)+ '/items/', data=dict(
-            title='Indian Ocean',
-            deadline='13/11/1994',
-            status='Incomplete'
-        ), headers=dict(Authorization=self.auth_token))
-        rv2 = self.app.get('/api/v1/bucketlists/' + str(bucketlistID), headers=dict(Authorization=self.auth_token))
-        data2 = json.loads(rv2.data.decode())
-        self.assertEqual(data2['1'], 'Indian Ocean')
-
-    def test_authenticated_user_can_get_search_and_paginate_items_in_bucketlist(self):
-        bucketlist = Bucketlist(name='Water bodies')
-        db.session.add(bucketlist)
-        self.user.bucketlists.append(bucketlist)  # FK relationship
-        db.session.commit()
-        bucketlistID = bucketlist.id
-        rv1 = self.app.get('/api/v1/bucketlists/' + str(bucketlistID) + '?limit=2',
-                           headers=dict(Authorization=self.auth_token))
-        data = json.loads(rv1.data.decode())
-        self.assertEqual(data['message'], 'No items in this bucketlist')
-        self.app.post('/api/v1/bucketlists/' + str(bucketlistID)+ '/items/', data=dict(
-            title='Indian Ocean',
-            deadline='13/11/1994',
-            status='Incomplete'
-        ), headers=dict(Authorization=self.auth_token))
-        self.app.post('/api/v1/bucketlists/' + str(bucketlistID) + '/items/', data=dict(
-            title='Pacific Ocean',
-            deadline='13/11/1994',
-            status='Incomplete'
-        ), headers=dict(Authorization=self.auth_token))
-        self.app.post('/api/v1/bucketlists/' + str(bucketlistID) + '/items/', data=dict(
-            title='Atlantic Ocean',
-            deadline='13/11/1994',
-            status='Incomplete'
-        ), headers=dict(Authorization=self.auth_token))
-        rv1= self.app.get('/api/v1/bucketlists/' + str(bucketlistID) + '?q=Ocean',
-                           headers=dict(Authorization=self.auth_token))
-        data = json.loads(rv1.data.decode())
-        self.assertEqual(len(data), 3)
-        rv1 = self.app.get('/api/v1/bucketlists/' + str(bucketlistID) + '?limit=2',
-                           headers=dict(Authorization=self.auth_token))
-        data = json.loads(rv1.data.decode())
-        self.assertEqual(len(data), 2)
-        rv2 = self.app.get('/api/v1/bucketlists/' + str(bucketlistID) + '?q=Zanzi',
-                           headers=dict(Authorization=self.auth_token))
-        data2 = json.loads(rv2.data.decode())
-        print(data2)
-        self.assertEqual(data2['message'], 'Bucketlist not found')
-        rv3 = self.app.get('/api/v1/bucketlists/' + str(bucketlistID) + '?q=Ocean&limit=2', headers=dict(Authorization=self.auth_token))
-        data3 = json.loads(rv3.data.decode())
-        self.assertEqual(len(data3), 2)
-        rv2 = self.app.get('/api/v1/bucketlists/' + str(bucketlistID) + '?q=Zanzi&limit=2', headers=dict(Authorization=self.auth_token))
-        data4 = json.loads(rv2.data.decode())
-        self.assertEqual(data4['message'], 'No result matches this search')
-
-    def test_user_can_update_item_in_bucketlist(self):
-        self.app.post('/api/v1/bucketlists/' + str(self.bucketlistID) + '/items/', data=dict(
-            title='Indian Ocean',
-            deadline='13/11/1994',
-            status='Incomplete'
-        ), headers=dict(Authorization=self.auth_token))
-        rv = self.app.put('/api/v1/bucketlists/' + str(self.bucketlistID) + '/items/1', data=dict(
-            title='Pacific Ocean',
-            deadline='13/11/1994',
-            status='Incomplete'
-        ), headers=dict(Authorization=self.auth_token))
-        data = json.loads(rv.data.decode())
-        self.assertEqual(data['message'], 'Bucketlist item updated successfully')
-
-    def test_user_cannot_update_item_in_bucketlist_that_doesnt_exist(self):
-        self.app.post('/api/v1/bucketlists/' + str(self.bucketlistID) + '/items/', data=dict(
-            title='Indian Ocean',
-            deadline='13/11/1994',
-            status='Incomplete'
-        ), headers=dict(Authorization=self.auth_token))
-        rv = self.app.put('/api/v1/bucketlists/' + str(100) + '/items/1', data=dict(
-            title='Pacific Ocean',
-            deadline='13/11/1994',
-            status='Incomplete'
-        ), headers=dict(Authorization=self.auth_token))
-        data = json.loads(rv.data.decode())
-        self.assertEqual(data['message'], 'Bucketlist item doesnt exist')
-
-    def test_unauthenticated_user_cannot_update_item_in_bucketlist(self):
-        self.app.post('/api/v1/bucketlists/' + str(self.bucketlistID) + '/items/', data=dict(
-            title='Indian Ocean',
-            deadline='13/11/1994',
-            status='Incomplete'
-        ), headers=dict(Authorization=self.auth_token))
-        rv = self.app.put('/api/v1/bucketlists/' + str(self.bucketlistID) + '/items/1', data=dict(
-            title='Pacific Ocean',
-            deadline='13/11/1994',
-            status='Incomplete'
-        ), headers=dict(Authorization=None))
-        data = json.loads(rv.data.decode())
-        self.assertEqual(data['message'], 'Invalid token, Login again')
-
-    def test_user_cannot_update_item_in_bucketlist_without_token(self):
-        self.app.post('/api/v1/bucketlists/' + str(self.bucketlistID) + '/items/', data=dict(
-            title='Indian Ocean',
-            deadline='13/11/1994',
-            status='Incomplete'
-        ), headers=dict(Authorization=self.auth_token))
-        rv = self.app.put('/api/v1/bucketlists/' + str(self.bucketlistID) + '/items/1', data=dict(
-            title='Pacific Ocean',
-            deadline='13/11/1994',
-            status='Incomplete'
-        ))
-        data = json.loads(rv.data.decode())
-        self.assertEqual(data['message'], 'Token not found, Login to get one')
-
-    def test_user_can_delete_item_in_bucketlist(self):
-        self.app.post('/api/v1/bucketlists/' + str(self.bucketlistID) + '/items/', data=dict(
-            title='Indian Ocean',
-            deadline='13/11/1994',
-            status='Incomplete'
-        ), headers=dict(Authorization=self.auth_token))
-        rv = self.app.delete('/api/v1/bucketlists/' + str(self.bucketlistID) + '/items/1',
-                             headers=dict(Authorization=self.auth_token))
-        data = json.loads(rv.data.decode())
-        self.assertEqual(data['message'], 'Bucketlist item deleted successfully')
-
-    def test_user_cannot_delete_item_in_bucketlist_that_doesnt_exist(self):
-        self.app.post('/api/v1/bucketlists/' + str(self.bucketlistID) + '/items/', data=dict(
-            title='Indian Ocean',
-            deadline='13/11/1994',
-            status='Incomplete'
-        ), headers=dict(Authorization=self.auth_token))
-        rv = self.app.put('/api/v1/bucketlists/' + str(100) + '/items/1', data=dict(
-            title='Pacific Ocean',
-            deadline='13/11/1994',
-            status='Incomplete'
-        ), headers=dict(Authorization=self.auth_token))
-        data = json.loads(rv.data.decode())
-        self.assertEqual(data['message'], 'Bucketlist item doesnt exist')
-
-    def test_unauthenticated_user_cannot_delete_item_in_bucketlist(self):
-        self.app.post('/api/v1/bucketlists/' + str(self.bucketlistID) + '/items/', data=dict(
-            title='Indian Ocean',
-            deadline='13/11/1994',
-            status='Incomplete'
-        ), headers=dict(Authorization=self.auth_token))
-        rv = self.app.delete('/api/v1/bucketlists/' + str(self.bucketlistID) + '/items/1', headers=dict(Authorization=None))
-        data = json.loads(rv.data.decode())
-        self.assertEqual(data['message'], 'Invalid token, Login again')
-
-    def test_user_cannot_delete_item_in_bucketlist_without_token(self):
-        self.app.post('/api/v1/bucketlists/' + str(self.bucketlistID) + '/items/', data=dict(
-            title='Indian Ocean',
-            deadline='13/11/1994',
-            status='Incomplete'
-        ), headers=dict(Authorization=self.auth_token))
-        rv = self.app.delete('/api/v1/bucketlists/' + str(self.bucketlistID) + '/items/1')
-        data = json.loads(rv.data.decode())
-        self.assertEqual(data['message'], 'Token not found, Login to get one')
-
-if __name__ == '__main__':
-    unittest.main()
